move filer, db implementations to separate file
